extract to _deliveries
